chore: Split proto generator interface and implementation
