+++ dozer-api/src/generator/protoc/generator/template/proto.tmpl
